requirements.txt
Django==5.1.2
djangorestframework==3.15.2
psycopg[binary]==3.2.1
python-dotenv==1.0.1
pytz==2024.1
paho-mqtt==2.1.0
requests==2.32.3
djangorestframework-simplejwt